+++ tools/measurement/analyze_benchmark_output.py
@@ -16,12 +16,19 @@ import json
 import logging
 import math
 import sys
-from enum import StrEnum
 from pathlib import Path
 from typing import Annotated, Any, cast
 
 import cyclopts
 import pandas as pd
+from measurement_tools.cli import LogFormat, configure_logging, log_bad_input
+from measurement_tools.stats import median_or_none as _median_or_none
+from measurement_tools.stats import none_if_nan as _none_if_nan
+from measurement_tools.stats import sum_int_or_zero as _sum_int_or_zero
+from measurement_tools.stats import sum_or_none as _sum_or_none
+from measurement_tools.stats import sum_or_zero as _sum_or_zero
+from measurement_tools.tables import AnalysisExportResult, ExportFormat, ModelTableSpec
+from measurement_tools.tables import write_analysis_tables as _write_analysis_table_specs
 from pydantic import BaseModel, Field, computed_field
 
 app = cyclopts.App(help=__doc__)
@@ -36,20 +43,6 @@ _SIGNATURE_DETAIL_FIELDS = {
     "end_position",
     "value_length",
 }
-
-
-class ExportFormat(StrEnum):
-    parquet = "parquet"
-    csv = "csv"
-    jsonl = "jsonl"
-
-
-class LogFormat(StrEnum):
-    plain = "plain"
-    json = "json"
-
-
-_log_format = LogFormat.plain
 
 
 class CaseAnalysisRow(BaseModel):
@@ -312,34 +305,6 @@ class BenchmarkOutputAnalysis(BaseModel):
     @property
     def model_usage_group_count(self) -> int:
         return len(self.model_usage_groups)
-
-
-class TableSummary(BaseModel):
-    table: str
-    rows: int
-    path: str
-
-
-class AnalysisExportResult(BaseModel):
-    output_dir: str
-    format: ExportFormat
-    tables: list[TableSummary]
-    manifest_path: str
-
-
-def configure_logging(log_format: LogFormat) -> None:
-    global _log_format
-
-    _log_format = log_format
-    logging.basicConfig(level=logging.INFO, format="%(levelname)s: %(message)s")
-
-
-def log_bad_input(error: str) -> None:
-    if _log_format == LogFormat.json:
-        payload = {"level": "error", "event": "bad_input", "error": error}
-        sys.stderr.write(json.dumps(payload, ensure_ascii=True, sort_keys=True) + "\n")
-        return
-    logger.error("bad_input error=%s", error)
 
 
 def analyze_benchmark_output(
@@ -1048,21 +1013,6 @@ def _signature_count(artifact_rows: pd.DataFrame, *, signature_hashes: list[str]
     return _sum_or_none(artifact_rows, "final_entity_signature_count")
 
 
-def _sum_or_zero(dataframe: pd.DataFrame, column: str) -> float:
-    if column not in dataframe.columns:
-        return 0.0
-    return float(pd.to_numeric(dataframe[column], errors="coerce").fillna(0).sum())
-
-
-def _sum_or_none(dataframe: pd.DataFrame, column: str) -> float | None:
-    if column not in dataframe.columns:
-        return None
-    values = pd.to_numeric(dataframe[column], errors="coerce").dropna()
-    if values.empty:
-        return None
-    return float(values.sum())
-
-
 def _positive_count(dataframe: pd.DataFrame, column: str) -> int:
     if column not in dataframe.columns:
         return 0
@@ -1349,12 +1299,6 @@ def _build_group_row(keys: tuple[Any, ...], group: pd.DataFrame) -> GroupAnalysi
     )
 
 
-def _none_if_nan(value: object) -> str | None:
-    if pd.isna(value):
-        return None
-    return str(value)
-
-
 def _int_if_not_nan(value: object) -> int | None:
     if pd.isna(value):
         return None
@@ -1365,17 +1309,6 @@ def _float_if_not_nan(value: object) -> float | None:
     if pd.isna(value):
         return None
     return float(cast(Any, value))
-
-
-def _median_or_none(dataframe: pd.DataFrame, column: str) -> float | None:
-    values = pd.to_numeric(dataframe[column], errors="coerce").dropna()
-    if values.empty:
-        return None
-    return float(values.median())
-
-
-def _sum_int_or_zero(dataframe: pd.DataFrame, column: str) -> int:
-    return int(_sum_or_zero(dataframe, column))
 
 
 def _sum_bool_or_zero(dataframe: pd.DataFrame, column: str) -> int:
@@ -1452,54 +1385,16 @@ def write_analysis_tables(
     output_dir: Path,
     export_format: ExportFormat,
 ) -> AnalysisExportResult:
-    output_dir.mkdir(parents=True, exist_ok=True)
-    tables = [
-        _write_model_rows(
-            result.cases, output_dir / f"case_analysis.{export_format.value}", export_format, CaseAnalysisRow
-        ),
-        _write_model_rows(
-            result.groups, output_dir / f"group_analysis.{export_format.value}", export_format, GroupAnalysisRow
-        ),
-        _write_model_rows(
-            result.model_usage,
-            output_dir / f"model_analysis.{export_format.value}",
-            export_format,
-            ModelUsageAnalysisRow,
-        ),
-        _write_model_rows(
-            result.model_usage_groups,
-            output_dir / f"model_group_analysis.{export_format.value}",
-            export_format,
-            ModelUsageGroupAnalysisRow,
-        ),
-    ]
-    export_result = AnalysisExportResult(
-        output_dir=str(output_dir),
-        format=export_format,
-        tables=tables,
-        manifest_path=str(output_dir / "manifest.json"),
+    return _write_analysis_table_specs(
+        output_dir,
+        export_format,
+        [
+            ModelTableSpec("case_analysis", result.cases, CaseAnalysisRow),
+            ModelTableSpec("group_analysis", result.groups, GroupAnalysisRow),
+            ModelTableSpec("model_analysis", result.model_usage, ModelUsageAnalysisRow),
+            ModelTableSpec("model_group_analysis", result.model_usage_groups, ModelUsageGroupAnalysisRow),
+        ],
     )
-    Path(export_result.manifest_path).write_text(export_result.model_dump_json(indent=2) + "\n", encoding="utf-8")
-    return export_result
-
-
-def _write_model_rows(
-    rows: list[BaseModel],
-    path: Path,
-    export_format: ExportFormat,
-    row_model: type[BaseModel],
-) -> TableSummary:
-    if rows:
-        table = pd.json_normalize([row.model_dump() for row in rows], sep=".")
-    else:
-        table = pd.DataFrame(columns=list(row_model.model_fields))
-    if export_format == ExportFormat.parquet:
-        table.to_parquet(path, index=False)
-    elif export_format == ExportFormat.csv:
-        table.to_csv(path, index=False)
-    else:
-        table.to_json(path, orient="records", lines=True)
-    return TableSummary(table=path.stem, rows=len(table), path=str(path))
 
 
 def render_result(result: BenchmarkOutputAnalysis, *, json_output: bool) -> str:
@@ -1543,7 +1438,7 @@ def main(
         if output is not None:
             write_analysis_tables(result, output, format)
     except ValueError as exc:
-        log_bad_input(str(exc))
+        log_bad_input(logger, str(exc))
         raise SystemExit(125) from exc
     sys.stdout.write(render_result(result, json_output=json_output) + "\n")
