+++ tools/measurement/measurement_tools/__init__.py
@@ -1,0 +1,4 @@
+#!/usr/bin/env python
+# SPDX-FileCopyrightText: Copyright (c) 2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+# SPDX-License-Identifier: Apache-2.0
+"""Shared support for measurement command-line tools."""
